Use unsigned int

+++ quantum/rgb_matrix_animations/pixelation_anim.h
@@ -19,7 +19,7 @@ RGB_MATRIX_EFFECT(PIXELATION)
 #   ifdef RGB_MATRIX_CUSTOM_EFFECT_IMPLS
 
 static bool PIXELATION(effect_params_t* params) {
-    void set_rgb(int i, effect_params_t* params, bool on) {
+    void set_rgb(uint8_t i, effect_params_t* params, bool on) {
         if (!HAS_ANY_FLAGS(g_led_config.flags[i], params->flags)) { return; }
         if (on) {
             HSV hsv = {random8(), add8(random8() >> 1, 127), rgb_matrix_config.hsv.v};
